requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
ebooklib==0.18
paddleocr==2.7.3
paddlepaddle==2.5.2
Pillow==10.1.0
aiofiles==23.2.1
pydantic==2.5.0
beautifulsoup4==4.12.2
numpy==1.24.3